+++ src/Iced/Attribute/OnInput.hs
@@ -1,4 +1,6 @@
+{-# LANGUAGE FunctionalDependencies #-}
+
 module Iced.Attribute.OnInput where
 
-class UseOnInput cb a where
-  onInput :: cb -> a
+class UseOnInput callback attribute | attribute -> callback where
+  onInput :: callback -> attribute