+++ src/Iced/Attribute/OnRelease.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE FunctionalDependencies #-}
+
+module Iced.Attribute.OnRelease where
+
+class UseOnRelease message attribute | attribute -> message where
+  onRelease :: message -> attribute