+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * Added canvas frame methods: fillText, pushTransform, popTransform, rotate, scale
 * Removed dependency on TemplateHaskell
 * Renamed Command to Task
+* Added mouseArea widget